Create Code
Honey is a natural sweetener that has been consumed for centuries. However, the quality of honey can vary depending on several factors such as the type of flower the bees collected the nectar from, the location of the beehive, and the processing method used. Therefore, it is important to assess the quality of honey before consumption or commercial use.

In this project, we will use machine learning techniques to predict the quality of honey based on its physicochemical properties. We will use a dataset containing information on the physicochemical properties of different types of honey to train a machine learning algorithm to predict the quality of honey.